CPUMonitoring/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.1.7
django-celery==3.2.2
django-cors-headers==2.4.1
django-rest-framework==0.1.0
djangorestframework==3.9.1
kombu==3.0.37
psutil==5.5.1
pytz==2018.9
